backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
itsdangerous
orjson

# For CORS
starlette

# Reuse existing stack
Pillow
moviepy
python-dotenv
requests
playwright
playwright-stealth
